Friday 12h00
Finally everything appears to be working. The process handlers only do stuff with
global flags, and there is only a single thread handling both SIGINT and SIGALRM.
This thread also is the one to decide when to write to a file and when to end
execution.
Still needs more testing.